Fix bugzilla bug 313069: [Automation][Regression]Background image was shrunk in postscript.

+++ engine/org.eclipse.birt.report.engine.emitter.postscript/src/org/eclipse/birt/report/engine/emitter/postscript/PostscriptWriter.java
@@ -477,8 +477,8 @@ public class PostscriptWriter
 				resolutionX = 96;
 				resolutionY = 96;
 			}
-			imageWidth = image.getWidth( ) / resolutionX * 72;
-			imageHeight = image.getHeight( ) / resolutionY * 72;
+			imageWidth = ( (float) image.getWidth( ) ) / resolutionX * 72;
+			imageHeight = ( (float) image.getHeight( ) ) / resolutionY * 72;
 		}
 
 		Position imageSize = new Position( imageWidth, imageHeight );
